+++ src/pages/CommunityDetail.tsx
@@ -2,18 +2,15 @@ import { useState, useEffect } from 'react'
 import { Link } from 'react-router-dom'
 import { useParams } from 'react-router-dom'
 import photo from '../assets/profile.png'
-import { AiOutlineLike } from 'react-icons/ai'
-import { GoLink } from 'react-icons/go'
-import { URLCopy } from '@utils/formatDate'
 import { fetchCommunityDetail } from '../api/community'
 import type { PostData } from '@customType/communityDetail'
 import CommentsInfiniteScroll from '@components/commnunityDetail/CommentsInfiniteScroll'
 import MarkdownEdit from '@components/commnunityDetail/MarkdownEdit'
+import LikeAndCopyButton from '@components/commnunityDetail/LikeAndCopyButton'
 
 export default function CommunityDetail() {
   const { id } = useParams()
   const [postData, setPostData] = useState<PostData | null>(null)
-  const [isLike, setIsLike] = useState(false)
   const [likeNum, setLikeNum] = useState(2)
 
   useEffect(() => {
@@ -25,11 +22,6 @@ export default function CommunityDetail() {
     }
     fetchPost()
   }, [id])
-
-  const handleClickLike = () => {
-    setLikeNum((prev) => (isLike ? prev - 1 : prev + 1))
-    setIsLike((prev) => !prev)
-  }
 
   if (!postData) return <div className="text-center mt-36">로딩 중...</div>
 
@@ -80,31 +72,7 @@ export default function CommunityDetail() {
         </div>
         <div className="flex flex-col gap-[24px] w-full">
           <div className="flex w-full justify-end gap-[12px] pb-[24px] border-b-[1px] border-[#cecece]">
-            <button
-              className="flex gap-[4px] items-center text-[#707070] border-[1px] border-[#cecece] py-[10px] px-[16px] rounded-[1000px] w-[62px] h-[38px] cursor-pointer"
-              onClick={handleClickLike}
-            >
-              <AiOutlineLike
-                className={`h-[18px] w-[18px] ${isLike ? 'text-[#6201e0]' : 'text-[#707070]'}`}
-              />
-              <div
-                className={`text-[12px] font-[500] ${isLike ? 'text-[#6201e0]' : 'text-[#707070]'}`}
-              >
-                {likeNum}
-              </div>
-            </button>
-            <button
-              className="flex gap-[4px] items-center text-[#707070] border-[1px] border-[#cecece] py-[10px] px-[5px] rounded-[1000px] hover:bg-[#ececec] w-[82px] h-[38px] cursor-pointer"
-              onClick={async () => {
-                const result = await URLCopy()
-                alert(
-                  `${result ? '복사가 완료되었습니다.' : '복사가 실패하였습니다.'}`
-                )
-              }}
-            >
-              <GoLink className="h-[18px] w-[18px]" />
-              <div className="text-[12px] font-[500]">공유하기</div>
-            </button>
+            <LikeAndCopyButton likeNum={likeNum} setLikeNum={setLikeNum} />
           </div>
           <CommentsInfiniteScroll />
         </div>
